refactor: move orderProgressBar mixed files to module
