+++ apps/cowswap-frontend/src/modules/ethFlow/pure/EthFlowModalContent/EthFlowModalBottomContent.tsx
@@ -2,10 +2,10 @@ import { useCallback, useMemo, useState } from 'react'
 
 import { Trans } from '@lingui/macro'
 
-import { ActivityStatus } from 'legacy/hooks/useRecentActivity'
-
 import { SimpleAccountDetails } from 'modules/account/containers/SimpleAccountDetails'
 import { TradeFormBlankButton } from 'modules/tradeFormValidation'
+
+import { ActivityStatus } from 'common/types/activity'
 
 import { EthFlowActions } from '../../containers/EthFlow/hooks/useEthFlowActions'
 import { EthFlowState } from '../../services/ethFlow/types'
