+++ apps/cowswap-frontend/src/modules/ethFlow/state/ethFlowContextAtom.ts
@@ -1,6 +1,6 @@
 import { atom } from 'jotai'
 
-import { ActivityStatus } from '../../../legacy/hooks/useRecentActivity'
+import { ActivityStatus } from 'common/types/activity'
 
 export interface EthFlowActionContext {
   isNeeded: boolean
