+++ apps/cowswap-frontend/src/modules/orderProgressBar/OrderProgressBar/steps/CircularCountdown.tsx
@@ -1,4 +1,4 @@
-import { PROGRESS_BAR_TIMER_DURATION } from 'common/hooks/orderProgressBar'
+import { PROGRESS_BAR_TIMER_DURATION } from 'modules/orderProgressBarMixed/orderProgressBar'
 
 import * as styledEl from './styled'
