+++ apps/cowswap-frontend/src/modules/orderProgressBarMixed/orderProgressBar/atoms.ts
