+++ apps/cowswap-frontend/src/modules/orderProgressBarMixed/orderProgressBar/index.ts
