+++ apps/cowswap-frontend/src/modules/orderProgressBarMixed/orderProgressBar/progressBarStateMachine.md
