+++ apps/cowswap-frontend/src/modules/orderProgressBarMixed/orderProgressBar/types.ts
